WebSockets/tests/submissions/Opera/constructor/013.html

<!doctype html>
<title>WebSockets: multiple WebSocket objects</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  // test that the events are fired as they should when opening 25 websockets and
  // sending a message on each and then closing when getting the message back
  var ws = [];
  var events = 0;
  for (var i = 0; i < 25; ++i) {
    ws[i] = new WebSocket(SCHEME_AND_DOMAIN+':8007/echo');
    ws[i].id = i;
    ws[i].onopen = function(e) {
      setTestTimeout(5000);
      debug(e.type+' '+this.id);
      events++;
      this.send(this.id);
      this.onopen = assertUnreached;
    }
    ws[i].onmessage = function(e) {
      debug(e.type+' '+this.id);
      events++;
      assert_equals(e.data, ''+this.id);
      this.close();
      this.onmessage = assertUnreached;
    }
    ws[i].onclose = function(e) {
      debug(e.type+' '+this.id);
      events++;
      if (events == 75) {
        end();
      }
      this.onclose = assertUnreached;
    }
    ws[i].onerror = assertUnreached;
  }
});
</script>

